# fix conv2d bug + faster conv 1d