Closes #2190 and mostly closes #2180 

Dispatches 1D conv to 2D when it makes sense for perf reasons.

Benchmarks from https://github.com/ml-explore/mlx/issues/2180

| | torch | mlx pre | mlx post
--- | ---- | ---- | ----
conv1d| 0.198 ms |  0.812 ms | 0.166 ms
conv_transpose1d |  0.505 ms | 1.030 ms | 0.829 ms